requirements.txt
cryptography==42.0.8
djangorestframework==3.15.2
kombu==5.4.0
python-decouple==3.8
Requests==2.32.3
Django==4.2.5
celery==5.2.3
psycopg==3.1.10
psycopg2-binary==2.9.6
redis==4.5.0
django-cors-headers==4.3.1
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3



